use button_to for join-game button
Previously was handling this with Stimulus, and manually performing the redirect in js so it could
post the `chosenTargets`. But that didn't work with flash messages, so now this just sets a hidden
value on the `button_to` form.

Rails has a way to set arrays in form values, but it seemed like more work than just joining the
`chosenTargets`, and splitting again on the backend (since none of the piece kinds will ever have
a `,`).